Move all k8s-related code out of pkg/skaffold/debug
