+++ pkg/skaffold/debug/apply_transforms.go
@@ -17,71 +17,21 @@ limitations under the License.
 package debug
 
 import (
-	"bufio"
-	"bytes"
 	"context"
 	"fmt"
 	"strings"
 
-	"github.com/sirupsen/logrus"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/serializer/json"
-	"k8s.io/client-go/kubernetes/scheme"
-
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/docker"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/graph"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/manifest"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/output/log"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner/runcontext"
 )
 
-var (
-	decodeFromYaml = scheme.Codecs.UniversalDeserializer().Decode
-	encodeAsYaml   = func(o runtime.Object) ([]byte, error) {
-		s := json.NewYAMLSerializer(json.DefaultMetaFactory, scheme.Scheme, scheme.Scheme)
-		var b bytes.Buffer
-		w := bufio.NewWriter(&b)
-		if err := s.Encode(o, w); err != nil {
-			return nil, err
-		}
-		w.Flush()
-		return b.Bytes(), nil
+var ConfigRetriever = func(ctx context.Context, image string, builds []graph.Artifact, registries map[string]bool) (ImageConfiguration, error) {
+	if artifact := findArtifact(image, builds); artifact != nil {
+		return retrieveImageConfiguration(ctx, artifact, registries)
 	}
-)
-
-// ApplyDebuggingTransforms applies language-platform-specific transforms to a list of manifests.
-func ApplyDebuggingTransforms(l manifest.ManifestList, builds []graph.Artifact, registries manifest.Registries) (manifest.ManifestList, error) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	retriever := func(image string) (imageConfiguration, error) {
-		if artifact := findArtifact(image, builds); artifact != nil {
-			return retrieveImageConfiguration(ctx, artifact, registries.InsecureRegistries)
-		}
-		return imageConfiguration{}, fmt.Errorf("no build artifact for %q", image)
-	}
-	return applyDebuggingTransforms(l, retriever, registries.DebugHelpersRegistry)
-}
-
-func applyDebuggingTransforms(l manifest.ManifestList, retriever configurationRetriever, debugHelpersRegistry string) (manifest.ManifestList, error) {
-	var updated manifest.ManifestList
-	for _, manifest := range l {
-		obj, _, err := decodeFromYaml(manifest, nil, nil)
-		if err != nil {
-			log.Entry(context.TODO()).Debugf("Unable to interpret manifest for debugging: %v\n", err)
-		} else if transformManifest(obj, retriever, debugHelpersRegistry) {
-			manifest, err = encodeAsYaml(obj)
-			if err != nil {
-				return nil, fmt.Errorf("marshalling yaml: %w", err)
-			}
-			if logrus.IsLevelEnabled(logrus.DebugLevel) {
-				log.Entry(context.TODO()).Debugln("Applied debugging transform:\n", string(manifest))
-			}
-		}
-		updated = append(updated, manifest)
-	}
-
-	return updated, nil
+	return ImageConfiguration{}, fmt.Errorf("no build artifact for %q", image)
 }
 
 // findArtifact finds the corresponding artifact for the given image.
@@ -102,32 +52,32 @@ func findArtifact(image string, builds []graph.Artifact) *graph.Artifact {
 
 // retrieveImageConfiguration retrieves the image container configuration for
 // the given build artifact
-func retrieveImageConfiguration(ctx context.Context, artifact *graph.Artifact, insecureRegistries map[string]bool) (imageConfiguration, error) {
+func retrieveImageConfiguration(ctx context.Context, artifact *graph.Artifact, insecureRegistries map[string]bool) (ImageConfiguration, error) {
 	// TODO: use the proper RunContext
 	apiClient, err := docker.NewAPIClient(ctx, &runcontext.RunContext{
 		InsecureRegistries: insecureRegistries,
 	})
 	if err != nil {
-		return imageConfiguration{}, fmt.Errorf("could not connect to local docker daemon: %w", err)
+		return ImageConfiguration{}, fmt.Errorf("could not connect to local docker daemon: %w", err)
 	}
 
 	// the apiClient will go to the remote registry if local docker daemon is not available
 	manifest, err := apiClient.ConfigFile(ctx, artifact.Tag)
 	if err != nil {
 		log.Entry(ctx).Debugf("Error retrieving image manifest for %v: %v", artifact.Tag, err)
-		return imageConfiguration{}, fmt.Errorf("retrieving image config for %q: %w", artifact.Tag, err)
+		return ImageConfiguration{}, fmt.Errorf("retrieving image config for %q: %w", artifact.Tag, err)
 	}
 
 	config := manifest.Config
 	log.Entry(ctx).Debugf("Retrieved local image configuration for %v: %v", artifact.Tag, config)
 	// need to duplicate slices as apiClient caches requests
-	return imageConfiguration{
-		artifact:   artifact.ImageName,
-		env:        envAsMap(config.Env),
-		entrypoint: dupArray(config.Entrypoint),
-		arguments:  dupArray(config.Cmd),
-		labels:     dupMap(config.Labels),
-		workingDir: config.WorkingDir,
+	return ImageConfiguration{
+		Artifact:   artifact.ImageName,
+		Env:        envAsMap(config.Env),
+		Entrypoint: dupArray(config.Entrypoint),
+		Arguments:  dupArray(config.Cmd),
+		Labels:     dupMap(config.Labels),
+		WorkingDir: config.WorkingDir,
 	}, nil
 }
