+++ pkg/skaffold/debug/transform.go
@@ -54,45 +54,36 @@ package debug
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"strings"
-	"time"
-
-	appsv1 "k8s.io/api/apps/v1"
-	batchv1 "k8s.io/api/batch/v1"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/debug/annotations"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/debug/types"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/output/log"
 )
 
-// portAllocator is a function that takes a desired port and returns an available port
-// Ports are normally uint16 but Kubernetes types.ContainerPort.types.ContainerPort is an integer
-type portAllocator func(int32) int32
+// PortAllocator is a function that takes a desired port and returns an available port
+// Ports are normally uint16 but Kubernetes types.ContainerPort is an integer
+type PortAllocator func(int32) int32
 
 // configurationRetriever retrieves an container image configuration
-type configurationRetriever func(string) (imageConfiguration, error)
+type ConfigurationRetriever func(string) (ImageConfiguration, error)
 
-// imageConfiguration captures information from a docker/oci image configuration.
+// ImageConfiguration captures information from a docker/oci image configuration.
 // It also includes a "artifact", usually containing the corresponding artifact's' image name from `skaffold.yaml`.
-type imageConfiguration struct {
-	// artifact is the corresponding artifact's image name (`pkg/skaffold/build.Artifact.ImageName`)
-	artifact string
+type ImageConfiguration struct {
+	// Artifact is the corresponding Artifact's image name (`pkg/skaffold/build.Artifact.ImageName`)
+	Artifact string
 
-	labels     map[string]string
-	env        map[string]string
-	entrypoint []string
-	arguments  []string
-	workingDir string
+	Labels     map[string]string
+	Env        map[string]string
+	Entrypoint []string
+	Arguments  []string
+	WorkingDir string
 }
 
 const (
-	// debuggingSupportVolume is the name of the volume used to hold language runtime debugging support files.
-	debuggingSupportFilesVolume = "debugging-support-files"
+	// DebuggingSupportVolume is the name of the volume used to hold language runtime debugging support files.
+	DebuggingSupportFilesVolume = "debugging-support-files"
 )
 
 // entrypointLaunchers is a list of known entrypoints that effectively just launches the container image's CMD
@@ -115,182 +106,12 @@ func isEntrypointLauncher(entrypoint []string) bool {
 	return false
 }
 
-// transformManifest attempts to configure a manifest for debugging.
-// Returns true if changed, false otherwise.
-func transformManifest(obj runtime.Object, retrieveImageConfiguration configurationRetriever, debugHelpersRegistry string) bool {
-	one := int32(1)
-	switch o := obj.(type) {
-	case *v1.Pod:
-		return transformPodSpec(&o.ObjectMeta, &o.Spec, retrieveImageConfiguration, debugHelpersRegistry)
-	case *v1.PodList:
-		changed := false
-		for i := range o.Items {
-			if transformPodSpec(&o.Items[i].ObjectMeta, &o.Items[i].Spec, retrieveImageConfiguration, debugHelpersRegistry) {
-				changed = true
-			}
-		}
-		return changed
-	case *v1.ReplicationController:
-		if o.Spec.Replicas != nil {
-			o.Spec.Replicas = &one
-		}
-		return transformPodSpec(&o.Spec.Template.ObjectMeta, &o.Spec.Template.Spec, retrieveImageConfiguration, debugHelpersRegistry)
-	case *appsv1.Deployment:
-		if o.Spec.Replicas != nil {
-			o.Spec.Replicas = &one
-		}
-		return transformPodSpec(&o.Spec.Template.ObjectMeta, &o.Spec.Template.Spec, retrieveImageConfiguration, debugHelpersRegistry)
-	case *appsv1.DaemonSet:
-		return transformPodSpec(&o.Spec.Template.ObjectMeta, &o.Spec.Template.Spec, retrieveImageConfiguration, debugHelpersRegistry)
-	case *appsv1.ReplicaSet:
-		if o.Spec.Replicas != nil {
-			o.Spec.Replicas = &one
-		}
-		return transformPodSpec(&o.Spec.Template.ObjectMeta, &o.Spec.Template.Spec, retrieveImageConfiguration, debugHelpersRegistry)
-	case *appsv1.StatefulSet:
-		if o.Spec.Replicas != nil {
-			o.Spec.Replicas = &one
-		}
-		return transformPodSpec(&o.Spec.Template.ObjectMeta, &o.Spec.Template.Spec, retrieveImageConfiguration, debugHelpersRegistry)
-	case *batchv1.Job:
-		return transformPodSpec(&o.Spec.Template.ObjectMeta, &o.Spec.Template.Spec, retrieveImageConfiguration, debugHelpersRegistry)
-
-	default:
-		group, version, _, description := describe(obj)
-		if group == "apps" || group == "batch" {
-			if version != "v1" {
-				// treat deprecated objects as errors
-				log.Entry(context.TODO()).Errorf("deprecated versions not supported by debug: %s (%s)", description, version)
-			} else {
-				log.Entry(context.TODO()).Warnf("no debug transformation for: %s", description)
-			}
-		} else {
-			log.Entry(context.TODO()).Debugf("no debug transformation for: %s", description)
-		}
-		return false
-	}
-}
-
-// transformPodSpec attempts to configure a podspec for debugging.
-// Returns true if changed, false otherwise.
-func transformPodSpec(metadata *metav1.ObjectMeta, podSpec *v1.PodSpec, retrieveImageConfiguration configurationRetriever, debugHelpersRegistry string) bool {
-	// order matters as rewriteProbes only affects containers marked for debugging
-	containers := rewriteContainers(metadata, podSpec, retrieveImageConfiguration, debugHelpersRegistry)
-	timeouts := rewriteProbes(metadata, podSpec)
-	return containers || timeouts
-}
-
-// rewriteProbes rewrites k8s probes to expand timeouts to 10 minutes to allow debugging local probes.
-func rewriteProbes(metadata *metav1.ObjectMeta, podSpec *v1.PodSpec) bool {
-	var minTimeout time.Duration = 10 * time.Minute // make it configurable?
-	if annotation, found := metadata.Annotations[annotations.DebugProbeTimeouts]; found {
-		if annotation == "skip" {
-			log.Entry(context.TODO()).Debugf("skipping probe rewrite on %q by request", metadata.Name)
-			return false
-		}
-		if d, err := time.ParseDuration(annotation); err != nil {
-			log.Entry(context.TODO()).Warnf("invalid probe timeout value for %q: %q: %v", metadata.Name, annotation, err)
-		} else {
-			minTimeout = d
-		}
-	}
-	annotation, found := metadata.Annotations[annotations.DebugConfig]
-	if !found {
-		log.Entry(context.TODO()).Debugf("skipping probe rewrite on %q: not configured for debugging", metadata.Name)
-		return false
-	}
-	var config map[string]annotations.ContainerDebugConfiguration
-	if err := json.Unmarshal([]byte(annotation), &config); err != nil {
-		log.Entry(context.TODO()).Warnf("error unmarshalling debugging configuration for %q: %v", metadata.Name, err)
-		return false
-	}
-
-	changed := false
-	for i := range podSpec.Containers {
-		c := &podSpec.Containers[i]
-		// only affect containers listed in debug-config
-		if _, found := config[c.Name]; found {
-			lp := rewriteHTTPGetProbe(c.LivenessProbe, minTimeout)
-			rp := rewriteHTTPGetProbe(c.ReadinessProbe, minTimeout)
-			sp := rewriteHTTPGetProbe(c.StartupProbe, minTimeout)
-			if lp || rp || sp {
-				log.Entry(context.TODO()).Infof("Updated probe timeouts for %s/%s", metadata.Name, c.Name)
-			}
-			changed = changed || lp || rp || sp
-		}
-	}
-	return changed
-}
-
-func rewriteHTTPGetProbe(probe *v1.Probe, minTimeout time.Duration) bool {
-	if probe == nil || probe.HTTPGet == nil || int32(minTimeout.Seconds()) < probe.TimeoutSeconds {
-		return false
-	}
-	probe.TimeoutSeconds = int32(minTimeout.Seconds())
-	return true
-}
-
-// allocatePort walks the podSpec's containers looking for an available port that is close to desiredPort.
-// We deal with wrapping and avoid allocating ports < 1024
-func allocatePort(podSpec *v1.PodSpec, desiredPort int32) int32 {
-	var maxPort int32 = 65535 // ports are normally [1-65535]
-	if desiredPort < 1024 || desiredPort > maxPort {
-		desiredPort = 1024 // skip reserved ports
-	}
-	// We assume ports are rather sparsely allocated, so even if desiredPort
-	// is allocated, desiredPort+1 or desiredPort+2 are likely to be free
-	for port := desiredPort; port < maxPort; port++ {
-		if isPortAvailable(podSpec, port) {
-			return port
-		}
-	}
-	for port := desiredPort; port > 1024; port-- {
-		if isPortAvailable(podSpec, port) {
-			return port
-		}
-	}
-	panic("cannot find available port") // exceedingly unlikely
-}
-
-// isPortAvailable returns true if none of the pod's containers specify the given port.
-func isPortAvailable(podSpec *v1.PodSpec, port int32) bool {
-	for _, container := range podSpec.Containers {
-		for _, portSpec := range container.Ports {
-			if portSpec.ContainerPort == port {
-				return false
-			}
-		}
-	}
-	return true
-}
-
-func encodeConfigurations(configurations map[string]annotations.ContainerDebugConfiguration) string {
+func EncodeConfigurations(configurations map[string]annotations.ContainerDebugConfiguration) string {
 	bytes, err := json.Marshal(configurations)
 	if err != nil {
 		return ""
 	}
 	return string(bytes)
-}
-
-func describe(obj runtime.Object) (group, version, kind, description string) {
-	// get metadata/name; shamelessly stolen from from k8s.io/cli-runtime/pkg/printers/name.go
-	name := "<unknown>"
-	if acc, err := meta.Accessor(obj); err == nil {
-		if n := acc.GetName(); len(n) > 0 {
-			name = n
-		}
-	}
-
-	gvk := obj.GetObjectKind().GroupVersionKind()
-	group = gvk.Group
-	version = gvk.Version
-	kind = gvk.Kind
-	if group == "" {
-		description = fmt.Sprintf("%s/%s", strings.ToLower(kind), name)
-	} else {
-		description = fmt.Sprintf("%s.%s/%s", strings.ToLower(kind), group, name)
-	}
-	return
 }
 
 // exposePort adds a `types.ContainerPort` instance or amends an existing entry with the same port.
