+++ pkg/skaffold/debug/transform_go_test.go
@@ -21,14 +21,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	appsv1 "k8s.io/api/apps/v1"
-	batchv1 "k8s.io/api/batch/v1"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/debug/annotations"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/debugging/adapter"
 	"github.com/GoogleContainerTools/skaffold/testutil"
 )
 
@@ -69,49 +62,49 @@ func TestExtractDlvArg(t *testing.T) {
 func TestDlvTransformer_IsApplicable(t *testing.T) {
 	tests := []struct {
 		description string
-		source      imageConfiguration
+		source      ImageConfiguration
 		launcher    string
 		result      bool
 	}{
 		{
 			description: "GOMAXPROCS",
-			source:      imageConfiguration{env: map[string]string{"GOMAXPROCS": "2"}},
+			source:      ImageConfiguration{Env: map[string]string{"GOMAXPROCS": "2"}},
 			result:      true,
 		},
 		{
 			description: "GOGC",
-			source:      imageConfiguration{env: map[string]string{"GOGC": "off"}},
+			source:      ImageConfiguration{Env: map[string]string{"GOGC": "off"}},
 			result:      true,
 		},
 		{
 			description: "GODEBUG",
-			source:      imageConfiguration{env: map[string]string{"GODEBUG": "efence=1"}},
+			source:      ImageConfiguration{Env: map[string]string{"GODEBUG": "efence=1"}},
 			result:      true,
 		},
 		{
 			description: "GOTRACEBACK",
-			source:      imageConfiguration{env: map[string]string{"GOTRACEBACK": "off"}},
+			source:      ImageConfiguration{Env: map[string]string{"GOTRACEBACK": "off"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint with dlv",
-			source:      imageConfiguration{entrypoint: []string{"dlv", "exec", "--headless"}},
+			source:      ImageConfiguration{Entrypoint: []string{"dlv", "exec", "--headless"}},
 			result:      true,
 		},
 		{
 			description: "launcher entrypoint",
-			source:      imageConfiguration{entrypoint: []string{"launcher"}, arguments: []string{"dlv", "exec", "--headless"}},
+			source:      ImageConfiguration{Entrypoint: []string{"launcher"}, Arguments: []string{"dlv", "exec", "--headless"}},
 			launcher:    "launcher",
 			result:      true,
 		},
 		{
 			description: "entrypoint /bin/sh",
-			source:      imageConfiguration{entrypoint: []string{"/bin/sh"}},
+			source:      ImageConfiguration{Entrypoint: []string{"/bin/sh"}},
 			result:      false,
 		},
 		{
 			description: "nothing",
-			source:      imageConfiguration{},
+			source:      ImageConfiguration{},
 			result:      false,
 		},
 	}
@@ -122,486 +115,6 @@ func TestDlvTransformer_IsApplicable(t *testing.T) {
 			result := dlvTransformer{}.IsApplicable(test.source)
 
 			t.CheckDeepEqual(test.result, result)
-		})
-	}
-}
-
-func TestDlvTransformerApply(t *testing.T) {
-	tests := []struct {
-		description   string
-		containerSpec v1.Container
-		configuration imageConfiguration
-		shouldErr     bool
-		result        v1.Container
-		debugConfig   annotations.ContainerDebugConfiguration
-		image         string
-	}{
-		{
-			description:   "empty",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{},
-			shouldErr:     true,
-		},
-		{
-			description:   "basic",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{entrypoint: []string{"app", "arg"}},
-			result: v1.Container{
-				Command: []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-				Ports:   []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "go", Ports: map[string]uint32{"dlv": 56268}},
-			image:       "go",
-		},
-		{
-			description: "existing port",
-			containerSpec: v1.Container{
-				Ports: []v1.ContainerPort{{Name: "http-server", ContainerPort: 8080}},
-			},
-			configuration: imageConfiguration{entrypoint: []string{"app", "arg"}},
-			result: v1.Container{
-				Command: []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-				Ports:   []v1.ContainerPort{{Name: "http-server", ContainerPort: 8080}, {Name: "dlv", ContainerPort: 56268}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "go", Ports: map[string]uint32{"dlv": 56268}},
-			image:       "go",
-		},
-		{
-			description: "existing dlv port",
-			containerSpec: v1.Container{
-				Ports: []v1.ContainerPort{{Name: "dlv", ContainerPort: 7896}},
-			},
-			configuration: imageConfiguration{entrypoint: []string{"app", "arg"}},
-			result: v1.Container{
-				Command: []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-				Ports:   []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "go", Ports: map[string]uint32{"dlv": 56268}},
-			image:       "go",
-		},
-		{
-			description:   "command not entrypoint",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{arguments: []string{"app", "arg"}},
-			result: v1.Container{
-				Args:  []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-				Ports: []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "go", Ports: map[string]uint32{"dlv": 56268}},
-			image:       "go",
-		},
-		{
-			description: "entrypoint with args in container spec",
-			containerSpec: v1.Container{
-				Args: []string{"arg1", "arg2"},
-			},
-			configuration: imageConfiguration{entrypoint: []string{"app"}},
-			result: v1.Container{
-				Command: []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--"},
-				Args:    []string{"arg1", "arg2"},
-				Ports:   []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "go", Ports: map[string]uint32{"dlv": 56268}},
-			image:       "go",
-		},
-	}
-	var identity portAllocator = func(port int32) int32 {
-		return port
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			adapter := adapter.NewAdapter(&test.containerSpec)
-			config, image, err := dlvTransformer{}.Apply(adapter, test.configuration, identity, nil)
-			adapter.Apply()
-
-			t.CheckError(test.shouldErr, err)
-			t.CheckDeepEqual(test.result, test.containerSpec)
-			t.CheckDeepEqual(test.debugConfig, config)
-			t.CheckDeepEqual(test.image, image)
-		})
-	}
-}
-
-func TestTransformManifestDelve(t *testing.T) {
-	int32p := func(x int32) *int32 { return &x }
-	tests := []struct {
-		description string
-		in          runtime.Object
-		transformed bool
-		out         runtime.Object
-	}{
-		{
-			"Pod with no transformable container",
-			&v1.Pod{
-				Spec: v1.PodSpec{Containers: []v1.Container{
-					{
-						Name:    "test",
-						Command: []string{"echo", "Hello World"},
-					},
-				}}},
-			false,
-			&v1.Pod{
-				Spec: v1.PodSpec{Containers: []v1.Container{
-					{
-						Name:    "test",
-						Command: []string{"echo", "Hello World"},
-					},
-				}}},
-		},
-		{
-			"Pod with Go container with GOMAXPROCS",
-			&v1.Pod{
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{{
-						Name:    "test",
-						Command: []string{"app", "arg"},
-						Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-					}},
-				}},
-			true,
-			&v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-				},
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{{
-						Name:         "test",
-						Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-						Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-						Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-						VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-					}},
-					InitContainers: []v1.Container{{
-						Name:         "install-go-debug-support",
-						Image:        "HELPERS/go",
-						VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-					}},
-					Volumes: []v1.Volume{{
-						Name:         "debugging-support-files",
-						VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-					}},
-				}},
-		},
-		{
-			"Deployment with Go container",
-			&appsv1.Deployment{
-				Spec: appsv1.DeploymentSpec{
-					Replicas: int32p(2),
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"app", "arg"},
-								Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-							}},
-						}}}},
-			true,
-			&appsv1.Deployment{
-				// ObjectMeta: metav1.ObjectMeta{
-				//  Labels: map[string]string{"debug.cloud.google.com/enabled": `yes`},
-				// },
-				Spec: appsv1.DeploymentSpec{
-					Replicas: int32p(1),
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-								Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-								Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-go-debug-support",
-								Image:        "HELPERS/go",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"ReplicaSet with Go container",
-			&appsv1.ReplicaSet{
-				Spec: appsv1.ReplicaSetSpec{
-					Replicas: int32p(2),
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"app", "arg"},
-								Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-							}},
-						}}}},
-			true,
-			&appsv1.ReplicaSet{
-				// ObjectMeta: metav1.ObjectMeta{
-				//	Labels: map[string]string{"debug.cloud.google.com/enabled": `yes`},
-				// },
-				Spec: appsv1.ReplicaSetSpec{
-					Replicas: int32p(1),
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-								Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-								Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-go-debug-support",
-								Image:        "HELPERS/go",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"StatefulSet with Go container",
-			&appsv1.StatefulSet{
-				Spec: appsv1.StatefulSetSpec{
-					Replicas: int32p(2),
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"app", "arg"},
-								Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-							}},
-						}}}},
-			true,
-			&appsv1.StatefulSet{
-				// ObjectMeta: metav1.ObjectMeta{
-				//	Labels: map[string]string{"debug.cloud.google.com/enabled": `yes`},
-				// },
-				Spec: appsv1.StatefulSetSpec{
-					Replicas: int32p(1),
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-								Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-								Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-go-debug-support",
-								Image:        "HELPERS/go",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"DaemonSet with Go container",
-			&appsv1.DaemonSet{
-				Spec: appsv1.DaemonSetSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"app", "arg"},
-								Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-							}},
-						}}}},
-			true,
-			&appsv1.DaemonSet{
-				// ObjectMeta: metav1.ObjectMeta{
-				//	Labels: map[string]string{"debug.cloud.google.com/enabled": `yes`},
-				// },
-				Spec: appsv1.DaemonSetSpec{
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-								Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-								Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-go-debug-support",
-								Image:        "HELPERS/go",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"Job with Go container",
-			&batchv1.Job{
-				Spec: batchv1.JobSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"app", "arg"},
-								Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-							}},
-						}}}},
-			true,
-			&batchv1.Job{
-				// ObjectMeta: metav1.ObjectMeta{
-				//	Labels: map[string]string{"debug.cloud.google.com/enabled": `yes`},
-				// },
-				Spec: batchv1.JobSpec{
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-								Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-								Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-go-debug-support",
-								Image:        "HELPERS/go",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"ReplicationController with Go container",
-			&v1.ReplicationController{
-				Spec: v1.ReplicationControllerSpec{
-					Replicas: int32p(2),
-					Template: &v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"app", "arg"},
-								Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-							}},
-						}}}},
-			true,
-			&v1.ReplicationController{
-				// ObjectMeta: metav1.ObjectMeta{
-				//	Labels: map[string]string{"debug.cloud.google.com/enabled": `yes`},
-				// },
-				Spec: v1.ReplicationControllerSpec{
-					Replicas: int32p(1),
-					Template: &v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-								Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-								Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-go-debug-support",
-								Image:        "HELPERS/go",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"PodList with Go and non-Go container",
-			&v1.PodList{
-				Items: []v1.Pod{
-					{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "echo",
-								Command: []string{"echo", "Hello World"},
-							}},
-						}},
-					{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"app", "arg"},
-								Env:     []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-							}},
-						}},
-				}},
-			true,
-			&v1.PodList{
-				Items: []v1.Pod{
-					{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "echo",
-								Command: []string{"echo", "Hello World"},
-							}},
-						}},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"go","ports":{"dlv":56268}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"/dbg/go/bin/dlv", "exec", "--headless", "--continue", "--accept-multiclient", "--listen=:56268", "--api-version=2", "app", "--", "arg"},
-								Ports:        []v1.ContainerPort{{Name: "dlv", ContainerPort: 56268}},
-								Env:          []v1.EnvVar{{Name: "GOMAXPROCS", Value: "1"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-go-debug-support",
-								Image:        "HELPERS/go",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}},
-				}},
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			value := test.in.DeepCopyObject()
-
-			retriever := func(image string) (imageConfiguration, error) {
-				return imageConfiguration{}, nil
-			}
-			result := transformManifest(value, retriever, "HELPERS")
-
-			t.CheckDeepEqual(test.transformed, result)
-			t.CheckDeepEqual(test.out, value)
 		})
 	}
 }
