+++ pkg/skaffold/debug/transform_nodejs_test.go
@@ -21,14 +21,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	appsv1 "k8s.io/api/apps/v1"
-	batchv1 "k8s.io/api/batch/v1"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/debug/annotations"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/debugging/adapter"
 	"github.com/GoogleContainerTools/skaffold/testutil"
 )
 
@@ -65,104 +58,104 @@ func TestExtractInspectArg(t *testing.T) {
 func TestNodeTransformer_IsApplicable(t *testing.T) {
 	tests := []struct {
 		description string
-		source      imageConfiguration
+		source      ImageConfiguration
 		launcher    string
 		result      bool
 	}{
 		{
 			description: "NODE_VERSION",
-			source:      imageConfiguration{env: map[string]string{"NODE_VERSION": "10"}},
+			source:      ImageConfiguration{Env: map[string]string{"NODE_VERSION": "10"}},
 			result:      true,
 		},
 		{
 			description: "NODEJS_VERSION",
-			source:      imageConfiguration{env: map[string]string{"NODEJS_VERSION": "12"}},
+			source:      ImageConfiguration{Env: map[string]string{"NODEJS_VERSION": "12"}},
 			result:      true,
 		},
 		{
 			description: "NODE_ENV",
-			source:      imageConfiguration{env: map[string]string{"NODE_ENV": "production"}},
+			source:      ImageConfiguration{Env: map[string]string{"NODE_ENV": "production"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint node",
-			source:      imageConfiguration{entrypoint: []string{"node", "init.js"}},
+			source:      ImageConfiguration{Entrypoint: []string{"node", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint /usr/bin/node",
-			source:      imageConfiguration{entrypoint: []string{"/usr/bin/node", "init.js"}},
+			source:      ImageConfiguration{Entrypoint: []string{"/usr/bin/node", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "no entrypoint, args node",
-			source:      imageConfiguration{arguments: []string{"node", "init.js"}},
+			source:      ImageConfiguration{Arguments: []string{"node", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "no entrypoint, arguments /usr/bin/node",
-			source:      imageConfiguration{arguments: []string{"/usr/bin/node", "init.js"}},
+			source:      ImageConfiguration{Arguments: []string{"/usr/bin/node", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint nodemon",
-			source:      imageConfiguration{entrypoint: []string{"nodemon", "init.js"}},
+			source:      ImageConfiguration{Entrypoint: []string{"nodemon", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint /usr/bin/nodemon",
-			source:      imageConfiguration{entrypoint: []string{"/usr/bin/nodemon", "init.js"}},
+			source:      ImageConfiguration{Entrypoint: []string{"/usr/bin/nodemon", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "no entrypoint, args nodemon",
-			source:      imageConfiguration{arguments: []string{"nodemon", "init.js"}},
+			source:      ImageConfiguration{Arguments: []string{"nodemon", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "no entrypoint, arguments /usr/bin/nodemon",
-			source:      imageConfiguration{arguments: []string{"/usr/bin/nodemon", "init.js"}},
+			source:      ImageConfiguration{Arguments: []string{"/usr/bin/nodemon", "init.js"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint npm",
-			source:      imageConfiguration{entrypoint: []string{"npm", "run", "dev"}},
+			source:      ImageConfiguration{Entrypoint: []string{"npm", "run", "dev"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint /usr/bin/npm",
-			source:      imageConfiguration{entrypoint: []string{"/usr/bin/npm", "run", "dev"}},
+			source:      ImageConfiguration{Entrypoint: []string{"/usr/bin/npm", "run", "dev"}},
 			result:      true,
 		},
 		{
 			description: "no entrypoint, args npm",
-			source:      imageConfiguration{arguments: []string{"npm", "run", "dev"}},
+			source:      ImageConfiguration{Arguments: []string{"npm", "run", "dev"}},
 			result:      true,
 		},
 		{
 			description: "no entrypoint, arguments npm",
-			source:      imageConfiguration{arguments: []string{"npm", "run", "dev"}},
+			source:      ImageConfiguration{Arguments: []string{"npm", "run", "dev"}},
 			result:      true,
 		},
 		{
 			description: "no entrypoint, arguments /usr/bin/npm",
-			source:      imageConfiguration{arguments: []string{"/usr/bin/npm", "run", "dev"}},
+			source:      ImageConfiguration{Arguments: []string{"/usr/bin/npm", "run", "dev"}},
 			result:      true,
 		},
 		{
 			description: "entrypoint /bin/sh",
-			source:      imageConfiguration{entrypoint: []string{"/bin/sh"}},
+			source:      ImageConfiguration{Entrypoint: []string{"/bin/sh"}},
 			result:      false,
 		},
 		{
 			description: "entrypoint launcher", // `node` image docker-entrypoint.sh"
-			source:      imageConfiguration{entrypoint: []string{"docker-entrypoint.sh"}, arguments: []string{"npm", "run", "dev"}},
+			source:      ImageConfiguration{Entrypoint: []string{"docker-entrypoint.sh"}, Arguments: []string{"npm", "run", "dev"}},
 			launcher:    "docker-entrypoint.sh",
 			result:      true,
 		},
 		{
 			description: "nothing",
-			source:      imageConfiguration{},
+			source:      ImageConfiguration{},
 			result:      false,
 		},
 	}
@@ -206,467 +199,6 @@ func TestRewriteNpmCommandLine(t *testing.T) {
 			result := rewriteNpmCommandLine(test.in, inspectSpec{port: 9226})
 
 			t.CheckDeepEqual(test.result, result)
-		})
-	}
-}
-
-func TestNodeTransformer_Apply(t *testing.T) {
-	// no shouldErr as Apply always succeeds
-	tests := []struct {
-		description   string
-		containerSpec v1.Container
-		configuration imageConfiguration
-		result        v1.Container
-		debugConfig   annotations.ContainerDebugConfiguration
-	}{
-		{
-			description:   "empty",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{},
-			result: v1.Container{
-				Env:   []v1.EnvVar{{Name: "NODE_OPTIONS", Value: "--inspect=0.0.0.0:9229"}, {Name: "PATH", Value: "/dbg/nodejs/bin"}},
-				Ports: []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-		{
-			description:   "entrypoint",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{entrypoint: []string{"node"}},
-			result: v1.Container{
-				Command: []string{"node", "--inspect=0.0.0.0:9229"},
-				Env:     []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-				Ports:   []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-		{
-			description:   "entrypoint with PATH",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{entrypoint: []string{"node"}, env: map[string]string{"PATH": "/usr/bin"}},
-			result: v1.Container{
-				Command: []string{"node", "--inspect=0.0.0.0:9229"},
-				Env:     []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin:/usr/bin"}},
-				Ports:   []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-		{
-			description: "existing port",
-			containerSpec: v1.Container{
-				Ports: []v1.ContainerPort{{Name: "http-server", ContainerPort: 8080}},
-			},
-			configuration: imageConfiguration{entrypoint: []string{"node"}},
-			result: v1.Container{
-				Command: []string{"node", "--inspect=0.0.0.0:9229"},
-				Env:     []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-				Ports:   []v1.ContainerPort{{Name: "http-server", ContainerPort: 8080}, {Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-		{
-			description: "existing devtools port",
-			containerSpec: v1.Container{
-				Ports: []v1.ContainerPort{{Name: "devtools", ContainerPort: 4444}},
-			},
-			configuration: imageConfiguration{entrypoint: []string{"node"}},
-			result: v1.Container{
-				Command: []string{"node", "--inspect=0.0.0.0:9229"},
-				Env:     []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-				Ports:   []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-		{
-			description:   "command not entrypoint",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{arguments: []string{"node"}},
-			result: v1.Container{
-				Args:  []string{"node", "--inspect=0.0.0.0:9229"},
-				Env:   []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-				Ports: []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-		{
-			description:   "docker-entrypoint (#3821)",
-			containerSpec: v1.Container{},
-			configuration: imageConfiguration{
-				env:        map[string]string{"NODE_VERSION": "10.12"},
-				entrypoint: []string{"docker-entrypoint.sh"},
-				arguments:  []string{"npm run script"}},
-			result: v1.Container{
-				Env:   []v1.EnvVar{{Name: "NODE_OPTIONS", Value: "--inspect=0.0.0.0:9229"}, {Name: "PATH", Value: "/dbg/nodejs/bin"}},
-				Ports: []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-		{
-			description:   "image environment not copied",
-			containerSpec: v1.Container{Env: []v1.EnvVar{{Name: "OTHER", Value: "VALUE"}}},
-			configuration: imageConfiguration{entrypoint: []string{"node"}, env: map[string]string{"RANDOM": "VALUE"}},
-			result: v1.Container{
-				Command: []string{"node", "--inspect=0.0.0.0:9229"},
-				Env:     []v1.EnvVar{{Name: "OTHER", Value: "VALUE"}, {Name: "PATH", Value: "/dbg/nodejs/bin"}},
-				Ports:   []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-			},
-			debugConfig: annotations.ContainerDebugConfiguration{Runtime: "nodejs", Ports: map[string]uint32{"devtools": 9229}},
-		},
-	}
-	var identity portAllocator = func(port int32) int32 {
-		return port
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			adapter := adapter.NewAdapter(&test.containerSpec)
-			config, image, err := nodeTransformer{}.Apply(adapter, test.configuration, identity, nil)
-			adapter.Apply()
-
-			// Apply never fails since there's always the option to set NODE_OPTIONS
-			t.CheckNil(err)
-			t.CheckDeepEqual(test.result, test.containerSpec)
-			t.CheckDeepEqual(test.debugConfig, config)
-			t.CheckDeepEqual("nodejs", image)
-		})
-	}
-}
-
-func TestTransformManifestNodeJS(t *testing.T) {
-	int32p := func(x int32) *int32 { return &x }
-	tests := []struct {
-		description string
-		in          runtime.Object
-		transformed bool
-		out         runtime.Object
-	}{
-		{
-			"Pod with no transformable container",
-			&v1.Pod{
-				Spec: v1.PodSpec{Containers: []v1.Container{{
-					Name:    "test",
-					Command: []string{"echo", "Hello World"},
-				}}}},
-			false,
-			&v1.Pod{
-				Spec: v1.PodSpec{Containers: []v1.Container{{
-					Name:    "test",
-					Command: []string{"echo", "Hello World"},
-				}}}},
-		},
-		{
-			"Pod with NodeJS container",
-			&v1.Pod{
-				Spec: v1.PodSpec{Containers: []v1.Container{{
-					Name:    "test",
-					Command: []string{"node", "foo.js"},
-				}}}},
-			true,
-			&v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-				},
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{{
-						Name:         "test",
-						Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-						Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-						Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-						VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-					}},
-					InitContainers: []v1.Container{{
-						Name:         "install-nodejs-debug-support",
-						Image:        "HELPERS/nodejs",
-						VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-					}},
-					Volumes: []v1.Volume{{
-						Name:         "debugging-support-files",
-						VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-					}},
-				}},
-		},
-		{
-			"Deployment with NodeJS container",
-			&appsv1.Deployment{
-				Spec: appsv1.DeploymentSpec{
-					Replicas: int32p(2),
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{Containers: []v1.Container{{
-							Name:    "test",
-							Command: []string{"node", "foo.js"},
-						}}}}}},
-			true,
-			&appsv1.Deployment{
-				Spec: appsv1.DeploymentSpec{
-					Replicas: int32p(1),
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-								Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-								Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-nodejs-debug-support",
-								Image:        "HELPERS/nodejs",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"ReplicaSet with NodeJS container",
-			&appsv1.ReplicaSet{
-				Spec: appsv1.ReplicaSetSpec{
-					Replicas: int32p(2),
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{Containers: []v1.Container{{
-							Name:    "test",
-							Command: []string{"node", "foo.js"},
-						}}}}}},
-			true,
-			&appsv1.ReplicaSet{
-				Spec: appsv1.ReplicaSetSpec{
-					Replicas: int32p(1),
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-								Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-								Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-nodejs-debug-support",
-								Image:        "HELPERS/nodejs",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"StatefulSet with NodeJS container",
-			&appsv1.StatefulSet{
-				Spec: appsv1.StatefulSetSpec{
-					Replicas: int32p(2),
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{Containers: []v1.Container{{
-							Name:    "test",
-							Command: []string{"node", "foo.js"},
-						}}}}}},
-			true,
-			&appsv1.StatefulSet{
-				Spec: appsv1.StatefulSetSpec{
-					Replicas: int32p(1),
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-								Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-								Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-nodejs-debug-support",
-								Image:        "HELPERS/nodejs",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"DaemonSet with NodeJS container",
-			&appsv1.DaemonSet{
-				Spec: appsv1.DaemonSetSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{Containers: []v1.Container{{
-							Name:    "test",
-							Command: []string{"node", "foo.js"},
-						}}}}}},
-			true,
-			&appsv1.DaemonSet{
-				Spec: appsv1.DaemonSetSpec{
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-								Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-								Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-nodejs-debug-support",
-								Image:        "HELPERS/nodejs",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"Job with NodeJS container",
-			&batchv1.Job{
-				Spec: batchv1.JobSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "test",
-								Command: []string{"node", "foo.js"},
-							}}}}}},
-			true,
-			&batchv1.Job{
-				Spec: batchv1.JobSpec{
-					Template: v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-								Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-								Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-nodejs-debug-support",
-								Image:        "HELPERS/nodejs",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"ReplicationController with NodeJS container",
-			&v1.ReplicationController{
-				Spec: v1.ReplicationControllerSpec{
-					Replicas: int32p(2),
-					Template: &v1.PodTemplateSpec{
-						Spec: v1.PodSpec{Containers: []v1.Container{
-							{
-								Name:    "test",
-								Command: []string{"node", "foo.js"},
-							},
-						}}}}},
-			true,
-			&v1.ReplicationController{
-				Spec: v1.ReplicationControllerSpec{
-					Replicas: int32p(1),
-					Template: &v1.PodTemplateSpec{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-								Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-								Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-nodejs-debug-support",
-								Image:        "HELPERS/nodejs",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}},
-						}}}},
-		},
-		{
-			"PodList with Java and non-Java container",
-			&v1.PodList{
-				Items: []v1.Pod{
-					{
-						Spec: v1.PodSpec{Containers: []v1.Container{{
-							Name:    "echo",
-							Command: []string{"echo", "Hello World"},
-						}}},
-					},
-					{
-						Spec: v1.PodSpec{Containers: []v1.Container{{
-							Name:    "test",
-							Command: []string{"node", "foo.js"},
-						}}},
-					},
-				}},
-			true,
-			&v1.PodList{
-				Items: []v1.Pod{
-					{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:    "echo",
-								Command: []string{"echo", "Hello World"},
-							}}},
-					},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Annotations: map[string]string{"debug.cloud.google.com/config": `{"test":{"runtime":"nodejs","ports":{"devtools":9229}}}`},
-						},
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Name:         "test",
-								Command:      []string{"node", "--inspect=0.0.0.0:9229", "foo.js"},
-								Ports:        []v1.ContainerPort{{Name: "devtools", ContainerPort: 9229}},
-								Env:          []v1.EnvVar{{Name: "PATH", Value: "/dbg/nodejs/bin"}},
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							InitContainers: []v1.Container{{
-								Name:         "install-nodejs-debug-support",
-								Image:        "HELPERS/nodejs",
-								VolumeMounts: []v1.VolumeMount{{Name: "debugging-support-files", MountPath: "/dbg"}},
-							}},
-							Volumes: []v1.Volume{{
-								Name:         "debugging-support-files",
-								VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}},
-							}}},
-					}}},
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			value := test.in.DeepCopyObject()
-
-			retriever := func(image string) (imageConfiguration, error) {
-				return imageConfiguration{}, nil
-			}
-			result := transformManifest(value, retriever, "HELPERS")
-
-			t.CheckDeepEqual(test.transformed, result)
-			t.CheckDeepEqual(test.out, value)
 		})
 	}
 }
